Schema: accept the documents the product actually writes (#25)
Two constraints were stricter than every implementation around them, so
validating a real resume failed and saving an edit through the API/connector
was impossible.

- styling.fontSize: integer/min 8 -> number/min 6. The builders' stepper
  moves in 0.25pt steps over [6,16] (DEFAULT_TEMPLATE_BOUNDS.fontSize in the
  document renderer), so documents legitimately hold 7.75.
- MonthDate: allow year-only (YYYY) alongside YYYY-MM. The AI parser emits
  it for sources that list only years, the web app's legacy->V1 mapper passes
  it through deliberately, and the renderer formats it as the bare year.

Both widen validation only: types stay string/number, so generated bindings
keep their shapes. Rationale lives in YAML comments rather than description
so it does not duplicate into 11 generated doc comments. Month 13 and
out-of-range sizes still reject (verified with ajv against the real failing
document).

Also regenerates kotlin inbox/BaseEvent.kt, which was left stale when
aggregationKey was declared in #23.

+++ go/resume/resume.gen.go
@@ -272,10 +272,10 @@ type ActivityItem struct {
 	// Description HTML allowed.
 	Description *string `json:"description,omitempty"`
 
-	// EndDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+	// EndDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 	EndDate *MonthDate `json:"endDate,omitempty"`
 
-	// StartDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+	// StartDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 	StartDate *MonthDate `json:"startDate,omitempty"`
 }
 
@@ -286,11 +286,11 @@ type CertificationItem struct {
 	// Description HTML allowed.
 	Description *string `json:"description,omitempty"`
 
-	// EndDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+	// EndDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 	EndDate     *MonthDate `json:"endDate,omitempty"`
 	Institution *string    `json:"institution,omitempty"`
 
-	// StartDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+	// StartDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 	StartDate *MonthDate `json:"startDate,omitempty"`
 }
 
@@ -341,12 +341,12 @@ type EducationItem struct {
 	// Description HTML allowed.
 	Description *string `json:"description,omitempty"`
 
-	// EndDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+	// EndDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 	EndDate     *MonthDate `json:"endDate,omitempty"`
 	Institution string     `json:"institution"`
 	Major       *string    `json:"major,omitempty"`
 
-	// StartDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+	// StartDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 	StartDate *MonthDate `json:"startDate,omitempty"`
 }
 
@@ -369,11 +369,11 @@ type EmploymentItem struct {
 	// Description HTML allowed (paragraphs and lists).
 	Description *string `json:"description,omitempty"`
 
-	// EndDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+	// EndDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 	EndDate  *MonthDate `json:"endDate,omitempty"`
 	Position string     `json:"position"`
 
-	// StartDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+	// StartDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 	StartDate *MonthDate `json:"startDate,omitempty"`
 }
 
@@ -417,7 +417,7 @@ type LanguagesSection struct {
 // LanguagesSectionType defines model for LanguagesSection.Type.
 type LanguagesSectionType string
 
-// MonthDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+// MonthDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 type MonthDate = string
 
 // PersonalSection defines model for PersonalSection.
@@ -449,10 +449,10 @@ type ProjectItem struct {
 	// Description HTML allowed.
 	Description *string `json:"description,omitempty"`
 
-	// EndDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+	// EndDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 	EndDate *MonthDate `json:"endDate,omitempty"`
 
-	// StartDate Month precision (YYYY-MM). Use null endDate + current=true for ongoing entries.
+	// StartDate Month precision (YYYY-MM), or year-only (YYYY) when the source states no month. Use null endDate + current=true for ongoing entries.
 	StartDate *MonthDate `json:"startDate,omitempty"`
 	Title     string     `json:"title"`
 
@@ -587,8 +587,8 @@ type Styling struct {
 	// Font Font family override. Falls back to the template's default.
 	Font *string `json:"font,omitempty"`
 
-	// FontSize Body font size in points.
-	FontSize *int `json:"fontSize,omitempty"`
+	// FontSize Body font size in points. Fractional values are valid (the builders step by 0.25).
+	FontSize *float32 `json:"fontSize,omitempty"`
 
 	// LineHeight Body line-height multiplier (e.g. 1.2, 1.4).
 	LineHeight *float32 `json:"lineHeight,omitempty"`
